requirements.txt
numpy
pandas
ipykernel
sentence_transformers
faiss-cpu
streamlit